refactor(plan_SERVERSPLIT slice 11): the element-interaction bodies leave server.py
The twelve `element-interaction` tool bodies move VERBATIM out of
embedded/server.py into embedded/tool_sections/element_interaction.py. Pure
move: the only edits are the dropped registration decorator (contract rule 2)
and the rewrite of the singleton/knob reads to `rt.<name>`, resolved at CALL
time against the one patchable home (contract rule 3). Fifty-one rewrites, every
one inside an expression; not one docstring, signature or comment byte changed.

With this slice server.py holds ZERO tool bodies. What is left is mcp, the
registry, the binding loop that registers all 94 of them once per execution of
this module body, app_lifespan, the four @mcp.resource handlers, the xpool-safe
gate, build_arg_parser and the __main__ block, plus the migration alias block
slice 12 deletes.

This section went last because execute_script is the only body that reads THREE
runtime knobs at once — rt._script_rejection_reason (and through it
MAX_USER_SCRIPT_BYTES), rt.EXECUTE_SCRIPT_TIMEOUT and rt._clamp_timeout. The
guard itself never moved: it has lived in tool_runtime since slice 0, so what
this slice proves is the LAST caller reaching it from a section module, at call
time, where tests/conftest.py's patched_server can still substitute it.
take_screenshot's two function-local imports (io and PIL.Image) are carried as
they stood rather than hoisted, which keeps Pillow off this module's import
graph at binding time.

Census re-run: ZERO test re-points, for the eleventh section running. There is
no `from ...embedded.server import <name>`, no `patch("...server.<tool>")`
string target and no inspect.getsource in tests/ or tools/;
tests/test_execute_script_guard.py already reads EXECUTE_SCRIPT_TIMEOUT /
MAX_USER_SCRIPT_BYTES / _script_rejection_reason from tool_runtime (slice 0
re-pointed it) and reaches the tool itself by attribute on server, which the
binding loop still satisfies. release_evidence.HOST_EXEC_SITES is unaffected:
execute_script runs JavaScript in the BROWSER and is already filed under
BROWSER_EXEC_TOOLS, not the host-exec inventory.

Thirteen imports server.py was the last consumer of are pruned: base64, os,
tempfile, datetime, Path, ToolError, _require_tab, and the tool_runtime aliases
CDP_OPERATION_TIMEOUT, EXECUTE_SCRIPT_TIMEOUT, _clamp_timeout,
_script_rejection_reason, dom_handler and response_handler. Five aliases remain,
all with a named reader: browser_manager / network_interceptor / debug_logger
for the resource handlers, app_lifespan and __main__; clone_storage and
_with_cdp_timeout only because tests/test_clone_storage.py and
tests/test_observability.py read them off `server`, and both carry a suppressed
keep-reason saying slice 12 re-points and deletes them. §5.5's
MIGRATION_ALIASES floor RATCHETS DOWN 11 -> 5.

server.py 986 -> 524 LOC; the GRANDFATHER cap ratchets DOWN to the measured
post-ruff-format actual and tests/source_scan.py's floor ratchets UP 12 -> 13 —
its FINAL value, server.py + tool_runtime.py + all eleven section modules.
tool_sections/element_interaction.py is 509 LOC, inside the 1000-LOC default
with no grandfather row (plan §6.2). SECTION_MODULES is now complete, so the
mid-migration note about SECTION_TOOLS listing the moved sections first is
updated to say the order has settled — and it has: the live key order is now the
canonical one.

Verified: wire surface IDENTICAL to tests/goldens/tool_surface.json (94 tools),
and `git diff c3aef6e -- tests/goldens/` is empty across both slices;
SECTION_TOOLS 94 with per-app [94, 94, 94]; `--list-sections` subprocess
Total: 94 tools; xpool gate re-measured against a REAL backend subprocess over
HTTP (throwaway port, HOME/USERPROFILE redirected at a temp dir) — control 94,
--xpool-safe 81, --disable-cdp-functions 81, XPOOL_SAFE_MODE=1 81, i.e. still
exactly thirteen removed by each, unchanged from the slice-9 baseline; unit lane
2320 passed / 1 skipped (2326 at slice 10, -6 for the pruned aliases);
real-Chrome test_e2e_hard_dom.py 4 passed, test_e2e_interaction.py 11 passed,
test_e2e_interaction_fidelity.py 7 passed and test_e2e_transport.py 1 passed,
one file at a time with 120 live Chrome processes counted first; budgets, ruff
format+check, suppression owners, vulture, ty (96 warnings, 0 errors — the new
module contributes none, its annotations are already fully parameterised) and
gen_release_contract --check all clean.
