+++ pyproject.toml
@@ -325,6 +325,21 @@ ban-relative-imports = "all"
 # ARG001 is reload_page's `ignore_cache`, ASYNC109 navigate's `timeout`,
 # PERF401 list_instances' append loop and E501 eleven docstring lines. Nothing
 # widened.)
+"src/stealth_chrome_devtools_mcp/embedded/tool_sections/element_interaction.py" = [
+    "A002", "ANN401", "ARG001", "ASYNC109", "ASYNC230", "ASYNC240", "B904",
+    "BLE001", "DTZ005", "E501", "PLR0913", "PLR2004", "PLW2901", "PTH108",
+    "PTH123", "TID251", "TRY300",
+]  # plan_SERVERSPLIT slice 11 (the element-interaction bodies moved VERBATIM
+# out of server.py — the LAST section to leave it, and the largest. Every code
+# is the exact subset of server.py's file-wide list the moved lines fire, and
+# every one is carried, not authored: A002 is take_screenshot's `format`
+# parameter and ARG001 its `full_page`, both WIRE surface a pure move may not
+# rename; ANN401/TID251 are query_elements' and select_option's `Any` params,
+# ASYNC109 the three `timeout` params, PLR0913 type_text's seven; the whole
+# ASYNC230/ASYNC240/DTZ005/PLR2004/PLW2901/PTH108/PTH123 cluster is
+# take_screenshot's Pillow-and-tempfile compression path; B904/BLE001/TRY300 are
+# execute_script's and select_option's ToolError envelopes; E501 eight docstring
+# lines. Carried verbatim, nothing widened.)
 "src/stealth_chrome_devtools_mcp/embedded/clone_storage.py" = [
     "ANN001", "ANN202", "ANN401", "ARG001", "ARG005", "ASYNC240",
     "BLE001", "DTZ005", "DTZ006", "E501", "PIE810", "PLR2004", "PTH100",
