+++ tests/test_tool_sections_contract.py
@@ -275,10 +275,14 @@ def test_the_alias_set_is_not_empty():
     pruned one — ``file_based_element_cloner``, ``cdp_element_cloner``,
     ``cdp_function_executor`` — leaving 14. Slice 10 pruned three at once
     (``display_context``, ``dynamic_hook_system``, ``in_memory_storage``),
-    leaving 11. Set to the measured actual, never padded, so the pin still
-    cannot vanish silently but a deliberate prune is not read as one.
+    leaving 11, and slice 11 pruned six more, leaving 5: the three the four
+    ``@mcp.resource`` handlers, ``app_lifespan`` and the ``__main__`` block
+    still read, plus ``clone_storage`` and ``_with_cdp_timeout``, which survive
+    only because a test reads them off ``server`` until slice 12 re-points it.
+    Set to the measured actual, never padded, so the pin still cannot vanish
+    silently but a deliberate prune is not read as one.
     """
-    assert len(MIGRATION_ALIASES) >= 11, MIGRATION_ALIASES
+    assert len(MIGRATION_ALIASES) >= 5, MIGRATION_ALIASES
 
 
 @pytest.mark.parametrize("name", MIGRATION_ALIASES)
